+++ NEWS.md
@@ -1,3 +1,5 @@
+# mlr3cluster (development version)
+
 # mlr3cluster 0.2.0
 
 * feat: `Mlr3Error` and `Mlr3Warning` classes for errors and warnings.