+++ NEWS.md
@@ -1,4 +1,4 @@
-# mlr3cluster (development version)
+# mlr3cluster 0.3.0
 
 * feat: Add CLARA clustering learner `clust.clara` from the cluster package.
 * feat: Add k-prototypes clustering learner `clust.kproto` from the clustMixType package.